pyproject.toml
[project]
name = "potluck"
version = "0.1.0"
description = "Personal Knowledge Database MCP server - Personalize your AI with Google Takeout and GDPR exports"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [{ name = "DoubleGremlin181" }]
keywords = ["mcp", "knowledge-base", "personal-data", "google-takeout", "gdpr"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
]

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.12",

    # Database
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.14.0",
    "pgvector>=0.3.0",

    # Task queue
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",

    # MCP
    "mcp>=1.0.0",

    # Image processing (lightweight)
    "pillow>=10.0.0",
    "imagehash>=4.3.0",
    "exifread>=3.0.0",

    # Utilities
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "python-dateutil>=2.9.0",
    "tqdm>=4.66.0",

    # CLI
    "typer>=0.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]
ml = [
    # Embeddings (requires torch)
    "sentence-transformers>=3.0.0",
    "torch>=2.0.0",
    # OCR
    "easyocr>=1.7.0",
    # Face recognition (requires cmake + dlib)
    "face-recognition>=1.3.0",
]
all = [
    "potluck[dev,ml]",
]

[project.scripts]
potluck = "potluck.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/potluck"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "e2e: End-to-end tests requiring Docker (use --run-e2e to enable)",
]
addopts = "-m 'not e2e'"
